lesson7/requirements-langchain.txt
langchain>=0.2.5
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.7

openai>=1.12.0
python-dotenv>=1.0.1

langchainhub>=0.1.21
langsmith>=0.1.52
